+++ src/librdag/convertto.cc
@@ -38,13 +38,13 @@ ConvertTo::convertToOGRealMatrix(OGIntegerScalar::Ptr thing) const
 OGRealMatrix::Ptr
 ConvertTo::convertToOGRealMatrix(OGRealDiagonalMatrix::Ptr thing) const
 {
-  int rows = thing->getRows();
-  int cols = thing->getCols();
-  int wlen = thing->getDatalen();
+  size_t rows = thing->getRows();
+  size_t cols = thing->getCols();
+  size_t wlen = thing->getDatalen();
   OGRealMatrix::Ptr ret = OGRealMatrix::create(new real8[rows*cols](),rows,cols, OWNER);
   real8 * diagdata = thing->getData();
   real8 * data = ret->getData();
-  for(int i=0;i<wlen;i++)
+  for(size_t i=0;i<wlen;i++)
   {
     data[i+i*rows]=diagdata[i];
   }
@@ -54,9 +54,9 @@ ConvertTo::convertToOGRealMatrix(OGRealDiagonalMatrix::Ptr thing) const
 OGRealMatrix::Ptr
 ConvertTo::convertToOGRealMatrix(OGLogicalMatrix::Ptr thing) const
 {
-  int rows = thing->getRows();
-  int cols = thing->getCols();
-  int wlen = thing->getDatalen();
+  size_t rows = thing->getRows();
+  size_t cols = thing->getCols();
+  size_t wlen = thing->getDatalen();
   OGRealMatrix::Ptr ret = OGRealMatrix::create(new real8[wlen](),rows,cols, OWNER);
   real8 * thedata = thing->getData();
   real8 * data = ret->getData();
@@ -69,17 +69,17 @@ OGRealMatrix::Ptr
 ConvertTo::convertToOGRealMatrix(OGRealSparseMatrix::Ptr thing) const
 {
 
-  int rows = thing->getRows();
-  int cols = thing->getCols();
-  int * colPtr = thing->getColPtr();
-  int * rowIdx = thing->getRowIdx();
+  size_t rows = thing->getRows();
+  size_t cols = thing->getCols();
+  int4 * colPtr = thing->getColPtr();
+  int4 * rowIdx = thing->getRowIdx();
   real8 * sparsedata = thing->getData();
 
   OGRealMatrix::Ptr ret = OGRealMatrix::create(new real8[rows*cols](),rows,cols, OWNER);
   real8 * data = ret->getData();
-  for (int ir = 0; ir < cols; ir++)
+  for (size_t ir = 0; ir < cols; ir++)
   {
-    for (int i = colPtr[ir]; i < colPtr[ir + 1]; i++)
+    for (int4 i = colPtr[ir]; i < colPtr[ir + 1]; i++)
     {
       data[rowIdx[i] + ir * rows] = sparsedata[i];
     }
@@ -117,13 +117,13 @@ ConvertTo::convertToOGComplexMatrix(OGComplexScalar::Ptr thing) const
 OGComplexMatrix::Ptr
 ConvertTo::convertToOGComplexMatrix(OGRealDiagonalMatrix::Ptr thing) const
 {
-  int rows = thing->getRows();
-  int cols = thing->getCols();
-  int wlen = thing->getDatalen();
+  size_t rows = thing->getRows();
+  size_t cols = thing->getCols();
+  size_t wlen = thing->getDatalen();
   OGComplexMatrix::Ptr ret = OGComplexMatrix::create(new complex16[rows*cols](),rows,cols, OWNER);
   real8 * diagdata = thing->getData();
   complex16 * data = ret->getData();
-  for(int i=0;i<wlen;i++)
+  for(size_t i=0;i<wlen;i++)
   {
     data[i+i*rows]=diagdata[i];
   }
@@ -133,13 +133,13 @@ ConvertTo::convertToOGComplexMatrix(OGRealDiagonalMatrix::Ptr thing) const
 OGComplexMatrix::Ptr
 ConvertTo::convertToOGComplexMatrix(OGComplexDiagonalMatrix::Ptr thing) const
 {
-  int rows = thing->getRows();
-  int cols = thing->getCols();
-  int wlen = thing->getDatalen();
+  size_t rows = thing->getRows();
+  size_t cols = thing->getCols();
+  size_t wlen = thing->getDatalen();
   OGComplexMatrix::Ptr ret = OGComplexMatrix::create(new complex16[rows*cols](),rows,cols, OWNER);
   complex16 * diagdata = thing->getData();
   complex16 * data = ret->getData();
-  for(int i=0;i<wlen;i++)
+  for(size_t i=0;i<wlen;i++)
   {
     data[i+i*rows]=diagdata[i];
   }
@@ -149,16 +149,16 @@ ConvertTo::convertToOGComplexMatrix(OGComplexDiagonalMatrix::Ptr thing) const
 OGComplexMatrix::Ptr
 ConvertTo::convertToOGComplexMatrix(OGRealSparseMatrix::Ptr thing) const
 {
-  int rows = thing->getRows();
-  int cols = thing->getCols();
-  int * colPtr = thing->getColPtr();
-  int * rowIdx = thing->getRowIdx();
+  size_t rows = thing->getRows();
+  size_t cols = thing->getCols();
+  int4 * colPtr = thing->getColPtr();
+  int4 * rowIdx = thing->getRowIdx();
   real8 * sparsedata = thing->getData();
   OGComplexMatrix::Ptr ret = OGComplexMatrix::create(new complex16[rows*cols](),rows,cols, OWNER);
   complex16 * data = ret->getData();
-  for (int ir = 0; ir < cols; ir++)
+  for (size_t ir = 0; ir < cols; ir++)
   {
-    for (int i = colPtr[ir]; i < colPtr[ir + 1]; i++)
+    for (int4 i = colPtr[ir]; i < colPtr[ir + 1]; i++)
     {
       data[rowIdx[i] + ir * rows] = sparsedata[i];
     }
@@ -169,16 +169,16 @@ ConvertTo::convertToOGComplexMatrix(OGRealSparseMatrix::Ptr thing) const
 OGComplexMatrix::Ptr
 ConvertTo::convertToOGComplexMatrix(OGComplexSparseMatrix::Ptr thing) const
 {
-  int rows = thing->getRows();
-  int cols = thing->getCols();
-  int * colPtr = thing->getColPtr();
-  int * rowIdx = thing->getRowIdx();
+  size_t rows = thing->getRows();
+  size_t cols = thing->getCols();
+  int4 * colPtr = thing->getColPtr();
+  int4 * rowIdx = thing->getRowIdx();
   complex16 * sparsedata = thing->getData();
   OGComplexMatrix::Ptr ret = OGComplexMatrix::create(new complex16[rows*cols](),rows,cols, OWNER);
   complex16 * data = ret->getData();
-  for (int ir = 0; ir < cols; ir++)
+  for (size_t ir = 0; ir < cols; ir++)
   {
-    for (int i = colPtr[ir]; i < colPtr[ir + 1]; i++)
+    for (int4 i = colPtr[ir]; i < colPtr[ir + 1]; i++)
     {
       data[rowIdx[i] + ir * rows] = sparsedata[i];
     }
@@ -189,13 +189,13 @@ ConvertTo::convertToOGComplexMatrix(OGComplexSparseMatrix::Ptr thing) const
 OGComplexMatrix::Ptr
 ConvertTo::convertToOGComplexMatrix(OGRealMatrix::Ptr thing) const
 {
-  int rows = thing->getRows();
-  int cols = thing->getCols();
-  int wlen = thing->getDatalen();
+  size_t rows = thing->getRows();
+  size_t cols = thing->getCols();
+  size_t wlen = thing->getDatalen();
   OGComplexMatrix::Ptr ret = OGComplexMatrix::create(new complex16[wlen](),rows,cols, OWNER);
   real8 * densedata = thing->getData();
   complex16 * data = ret->getData();
-  for(int i=0;i<wlen;i++)
+  for(size_t i=0;i<wlen;i++)
   {
     data[i]=densedata[i];
   }
@@ -205,13 +205,13 @@ ConvertTo::convertToOGComplexMatrix(OGRealMatrix::Ptr thing) const
 OGComplexMatrix::Ptr
 ConvertTo::convertToOGComplexMatrix(OGLogicalMatrix::Ptr thing) const
 {
-  int rows = thing->getRows();
-  int cols = thing->getCols();
-  int wlen = thing->getDatalen();
+  size_t rows = thing->getRows();
+  size_t cols = thing->getCols();
+  size_t wlen = thing->getDatalen();
   OGComplexMatrix::Ptr ret = OGComplexMatrix::create(new complex16[wlen](),rows,cols, OWNER);
   real8 * densedata = thing->getData();
   complex16 * data = ret->getData();
-  for(int i=0;i<wlen;i++)
+  for(size_t i=0;i<wlen;i++)
   {
     data[i]=densedata[i];
   }
